match on vax1_date
This is done to avoid having to carry time_since_vax1 around unnecessarily.

Also vax1_date is consistent when the index_date for the control is updated from "potential" to "actual".

time_since_vax1 is now calculated at the end of the final extract (after matching), so it can be used in table 1